refactor
put class in its own classfile
