+++ Android/app/src/main/java/live/dittolive/skyservice/SkyServiceApplication.kt
@@ -62,19 +62,3 @@ class SkyServiceApplication: Application() {
         notificationManager.createNotificationChannel(channel)
     }
 }
-class AuthCallback: DittoAuthenticationCallback {
-    override fun authenticationRequired(authenticator: DittoAuthenticator) {
-        authenticator.loginWithToken(BuildConfig.DITTO_AUTH_TOKEN, BuildConfig.DITTO_AUTH_PROVIDER) { err ->
-            println("Login request completed. Error? $err")
-        }
-    }
-
-    override fun authenticationExpiringSoon(
-        authenticator: DittoAuthenticator,
-        secondsRemaining: Long
-    ) {
-        authenticator.loginWithToken(BuildConfig.DITTO_AUTH_TOKEN, BuildConfig.DITTO_AUTH_PROVIDER) { err ->
-            println("Login request completed. Error? $err")
-        }
-    }
-}
